Implement millisecond level indicator statistics

+++ store/src/main/java/org/apache/rocketmq/store/timer/rocksdb/TimerMessageKVStore.java
@@ -86,5 +86,5 @@ public interface TimerMessageKVStore {
      * @param key the key of the metric.
      * @param update the value of the metric.
      */
-    void syncMetric(int key, int update);
+    void syncMetric(long key, long update);
 }
